move to preexisting utils module
