+++ crates/ty_python_semantic/src/types/diagnostic.rs
@@ -1,8 +1,12 @@
 use super::call::CallErrorKind;
 use super::context::InferContext;
 use super::mro::DuplicateBaseError;
-use super::{CallArgumentTypes, CallDunderError, ClassBase, ClassLiteral, KnownClass};
+use super::{
+    CallArgumentTypes, CallDunderError, ClassBase, ClassLiteral, KnownClass,
+    add_inferred_python_version_hint_to_diagnostic,
+};
 use crate::db::Db;
+use crate::declare_lint;
 use crate::lint::{Level, LintRegistryBuilder, LintStatus};
 use crate::suppression::FileSuppressionId;
 use crate::types::LintDiagnosticGuard;
@@ -12,10 +16,8 @@ use crate::types::string_annotation::{
     RAW_STRING_TYPE_ANNOTATION,
 };
 use crate::types::{KnownFunction, KnownInstanceType, Type, protocol_class::ProtocolClassLiteral};
-use crate::{Program, PythonVersionWithSource, declare_lint};
 use itertools::Itertools;
-use ruff_db::diagnostic::{Annotation, Diagnostic, Severity, Span, SubDiagnostic};
-use ruff_db::files::system_path_to_file;
+use ruff_db::diagnostic::{Annotation, Diagnostic, Severity, SubDiagnostic};
 use ruff_python_ast::{self as ast, AnyNodeRef};
 use ruff_python_stdlib::builtins::version_builtin_was_added;
 use ruff_text_size::{Ranged, TextRange};
@@ -1760,52 +1762,6 @@ pub(super) fn report_possibly_unbound_attribute(
         "Attribute `{attribute}` on type `{}` is possibly unbound",
         object_ty.display(context.db()),
     ));
-}
-
-/// Add a subdiagnostic to `diagnostic` that explains why a certain Python version was inferred.
-///
-/// ty can infer the Python version from various sources, such as command-line arguments,
-/// configuration files, or defaults.
-pub fn add_inferred_python_version_hint_to_diagnostic(
-    db: &dyn Db,
-    diagnostic: &mut Diagnostic,
-    action: &str,
-) {
-    let program = Program::get(db);
-    let PythonVersionWithSource { version, source } = program.python_version_with_source(db);
-
-    match source {
-        crate::PythonVersionSource::Cli => {
-            diagnostic.info(format_args!(
-                "Python {version} was assumed when {action} because it was specified on the command line",
-            ));
-        }
-        crate::PythonVersionSource::File(path, range) => {
-            if let Ok(file) = system_path_to_file(db.upcast(), &**path) {
-                let mut sub_diagnostic = SubDiagnostic::new(
-                    Severity::Info,
-                    format_args!("Python {version} was assumed when {action}"),
-                );
-                sub_diagnostic.annotate(
-                    Annotation::primary(Span::from(file).with_optional_range(*range)).message(
-                        format_args!("Python {version} assumed due to this configuration setting"),
-                    ),
-                );
-                diagnostic.sub(sub_diagnostic);
-            } else {
-                diagnostic.info(format_args!(
-                    "Python {version} was assumed when {action} because of your configuration file(s)",
-                ));
-            }
-        }
-        crate::PythonVersionSource::Default => {
-            diagnostic.info(format_args!(
-                "Python {version} was assumed when {action} \
-                because it is the newest Python version supported by ty, \
-                and neither a command-line argument nor a configuration setting was provided",
-            ));
-        }
-    }
 }
 
 pub(super) fn report_unresolved_reference(context: &InferContext, expr_name_node: &ast::ExprName) {
