+++ crates/ty_python_semantic/src/util/mod.rs
@@ -1,1 +1,2 @@
+pub(crate) mod diagnostics;
 pub(crate) mod subscript;
